requirements.txt
requests
google-cloud-speech
glob